Hanh

+++ EHM/EHM_API/Repositories/OrderRepository.cs
@@ -195,7 +195,7 @@ public class OrderRepository : IOrderRepository
 			InvoiceId = o.InvoiceId,
 			TotalAmount = o.TotalAmount,
 			GuestPhone = o.GuestPhone,
-			Deposits = o.Deposits,
+			Deposits = (decimal)o.Deposits,
 			AddressId = o.AddressId ?? 0,
 			GuestAddress = o.Address?.GuestAddress,
 			ConsigneeName = o.Address?.ConsigneeName,
